Horário dinámico, ajustes no layout, controle de cidades já existentes

+++ main.js
@@ -1,7 +1,6 @@
 let result = document.querySelector('#result')
 let logo = document.querySelector('#logo')
 let cidade = document.querySelector('#cidade')
-let observacao = document.querySelector('#observacao')
 let insere = document.querySelector('#insere')
 let btnAPI = document.querySelector('#btnAPI')
 
@@ -11,8 +10,9 @@ let clear = document.querySelector('#clear')
 let controlers = document.querySelector('#controlers')
 
 window.addEventListener('load', listar)
-clear.addEventListener('click', limpar)
+//clear.addEventListener('click', limpar)
 insere.addEventListener('click', inserir)
+
 cidade.addEventListener('keyup', (e)=>{
     if(e.keyCode ==13){
         inserir()
@@ -26,6 +26,7 @@ let data = [];
  function inserir(){
 
     editaControl=false;
+
     if(cidade.value === '' || cidade.value.length < 2){ 
         alert(`The city is empty or its name is too short!`)
         limpaCampos()
@@ -37,9 +38,19 @@ let data = [];
 
     if(localStorage.getItem('data')){
         datas = JSON.parse(localStorage.getItem('data'))  
+        
+        let cidadeJaExiste = datas.filter((elem)=>{
+          return cidade.value == elem.cidade
+        })
+
+        if(cidadeJaExiste.length > 0){
+            alert(`Cidade ${cidade.value} já está cadastrada!`)
+            limpaCampos()
+            return}
+
+
         datas.unshift({
-            'cidade': cidade.value,
-            'observacao': observacao.value
+            'cidade': cidade.value
         })
         limpaCampos()
         localStorage.setItem('data', JSON.stringify(datas))
@@ -47,8 +58,7 @@ let data = [];
         }
     else{
         datas.unshift({
-            'cidade': cidade.value,
-            'observacao': observacao.value
+            'cidade': cidade.value
         })
         limpaCampos()
         localStorage.setItem('data', JSON.stringify(datas))
@@ -59,6 +69,7 @@ let data = [];
 
 //Listar 
 function listar(){
+
     result.innerHTML = '';
     limpaCampos()
     if(!localStorage.getItem('data')){
@@ -66,42 +77,40 @@ function listar(){
     }else{
         datas = JSON.parse(localStorage.getItem('data'))
         
+        //ITERAÇÂO 
         datas.forEach((element, index)=>{
 
-        if(KEY)
-        {    
-        api(element.cidade,index)
+        if(KEY){    
+        api(element.cidade,index).then((resp)=>{
+            element.fuso = resp;
+        })
         }
         
 
         result.insertAdjacentHTML('beforeend', `
-        <div id="card${index}" class="card">
+        <div id="card${index}" class="${KEY?"card":"card-show"}">
         
             <div id="inner-card-up"> 
+            <img id="icon${index}" src="">
                 <h3 id="cidade_card${index}">${element.cidade}</h3>
-                <img id="icon${index}" src="">
                 <h4 id="temperatura${index}"></h4>
             </div>
             
-            <div class="inner-card-medium" id="inner-card-medium${index}">
-            <p id="detalhes${index}"></p>
-            </div>   
-            
-            <div id="inner-card-medium">
-            <p id="observacoes_card">${element.observacao}</p>
-            </div>  
+             
 
             <div id="inner-card-down">
                 <h4 id="hora${index}"></h4>
                 <div>
-                <span Onclick=excluir(${index}) style="font-size: 25px"><i class="fa fa-times-circle"></i></span>
-                <span id="btnEditar${index}" Onclick=editar(${index}) style="font-size: 25px"><i  id="iconEditar${index}"class="fa fa-edit"></i></span>
+                <span Onclick=excluir(${index}) style="font-size: 20px"><i class="fa fa-times-circle"></i></span>
+                <span id="btnEditar${index}" Onclick=editar(${index}) style="font-size: 20px"><i  id="iconEditar${index}"class="fa fa-edit"></i></span>
                </div>
             </div>
         
         </div>`)
 
         })
+
+         setInterval(atualizaHora, 1000)
     }
 }
 
@@ -112,11 +121,11 @@ function editar(e){
 
     datas = JSON.parse(localStorage.getItem('data'))
 
+
     if(!editaControl){ 
     
     editaControl=true;
     cidade.value = datas[e].cidade;
-    observacao.value = datas[e].observacao;
     document.querySelector(`#iconEditar${e}`).className ='fa fa-check-circle';
     }else{
         if(cidade.value === '' || cidade.value.length < 2){ 
@@ -127,7 +136,6 @@ function editar(e){
         } 
         editaControl=false;
         datas[e].cidade = cidade.value;
-        datas[e].observacao = observacao.value;
         document.querySelector(`#iconEditar${e}`).className ='fa fa-edit';
         localStorage.setItem('data', JSON.stringify(datas))
         limpaCampos()
@@ -150,21 +158,43 @@ function limpar(){
 
 function limpaCampos(){
     cidade.value = "";
-    observacao.value = "";
-
 }
+
+
+
+function atualizaHora(){
+    if(datas.length < 1){return}
+    datas.forEach((elem, index)=>{
+        if(elem.fuso){ 
+            let novaHora = new Date((Math.floor(new Date().getTime()/1000)+(elem.fuso))*1000).toUTCString()
+            document.querySelector(`#hora${index}`).innerHTML = `${novaHora.slice(17,25)}`;
+        }
+        
+    })
+}
+
+
 
 
 //Chamando API 
  let dados,hora;
 
- function api(city, index){
-    fetch(`https://api.openweathermap.org/data/2.5/forecast?q=${city}&appid=48d5c83ce5c68b72bcd7900e669b8806&lang=pt_br&units=metric`)
+ async function api (city, index){
+    let fuso;
+
+    await fetch(`https://api.openweathermap.org/data/2.5/forecast?q=${city}&appid=48d5c83ce5c68b72bcd7900e669b8806&lang=pt_br&units=metric`)
     .then((res)=>{
        return res.json()
     }).then((json) =>{
         dados = json;
-        hora = new Date((Math.floor(new Date().getTime()/1000)+(dados.city.timezone))*1000).toUTCString()
+
+   
+            hora = new Date((Math.floor(new Date().getTime()/1000)+(dados.city.timezone))*1000).toUTCString()
+            
+            document.querySelector(`#hora${index}`).innerHTML = ` ${hora.slice(17,25)}`;
+
+        
+
         console.log(`
         Cidade: ${dados.city.name}
         País: ${dados.city.country}
@@ -177,17 +207,19 @@ function limpaCampos(){
 
         document.querySelector(`#temperatura${index}`).innerHTML = `${dados.list[0].main.temp}°C`;
         document.querySelector(`#cidade_card${index}`).innerHTML += `, ${dados.city.country}`;
-        document.querySelector(`#detalhes${index}`).innerHTML = `${dados.list[0].weather[0].description.toUpperCase()}.`;
-        document.querySelector(`#hora${index}`).innerHTML = ` ${hora.slice(17,25)}`;
+
+       
         document.querySelector(`#card${index}`).className = "card-show";
         document.querySelector(`#icon${index}`).src = `http://openweathermap.org/img/wn/${dados.list[0].weather[0].icon}.png`;  
 
-
+        fuso = dados.city.timezone;
 
     }).catch((err)=>{
         console.log(err)
         document.querySelector(`#card${index}`).className = "card-show";
     })
-    
+     return fuso;
        
 }
+
+
